requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.4
zipp==3.15.0
black

